update isort config, move enums
