+++ .pre-commit-config.yaml
@@ -15,11 +15,11 @@ repos:
         args:
           - --fix=lf
   - repo: https://github.com/psf/black-pre-commit-mirror
-    rev: 24.10.0
+    rev: 25.1.0
     hooks:
       - id: black
   - repo: https://github.com/pycqa/isort
-    rev: 5.13.2
+    rev: 6.0.1
     hooks:
       - id: isort
         args:
@@ -27,7 +27,7 @@ repos:
           - black
           - --filter-files
   - repo: https://github.com/RobertCraigie/pyright-python
-    rev: v1.1.390
+    rev: v1.1.399
     hooks:
       - id: pyright
         language_version: python3.13
