+++ ballsdex/core/utils/transformers.py
@@ -5,6 +5,13 @@ from enum import Enum
 from typing import TYPE_CHECKING, Generic, Iterable, Optional, TypeVar
 
 import discord
+from discord import app_commands
+from discord.interactions import Interaction
+from django.db.models import Model, Q
+from django.db.models.expressions import RawSQL
+from django.utils import timezone
+
+from ballsdex.settings import settings
 from bd_models.models import (
     Ball,
     BallInstance,
@@ -15,13 +22,6 @@ from bd_models.models import (
     economies,
     regimes,
 )
-from discord import app_commands
-from discord.interactions import Interaction
-from django.db.models import Model, Q
-from django.db.models.expressions import RawSQL
-from django.utils import timezone
-
-from ballsdex.settings import settings
 
 if TYPE_CHECKING:
     from ballsdex.core.bot import BallsDexBot
