+++ ballsdex/packages/countryballs/cog.py
@@ -3,12 +3,12 @@ import logging
 from typing import TYPE_CHECKING, cast
 
 import discord
-from bd_models.models import GuildConfig
 from discord.ext import commands
 
 from ballsdex.packages.countryballs.countryball import BallSpawnView
 from ballsdex.packages.countryballs.spawn import BaseSpawnManager
 from ballsdex.settings import settings
+from bd_models.models import GuildConfig
 
 if TYPE_CHECKING:
     from ballsdex.core.bot import BallsDexBot
