+++ ballsdex/packages/trade/display.py
@@ -1,11 +1,11 @@
 from typing import TYPE_CHECKING, Iterable
 
 import discord
-from bd_models.models import Trade as TradeModel
 
 from ballsdex.core.utils import menus
 from ballsdex.core.utils.paginator import Pages
 from ballsdex.packages.trade.trade_user import TradingUser
+from bd_models.models import Trade as TradeModel
 
 if TYPE_CHECKING:
     from ballsdex.core.bot import BallsDexBot
@@ -146,21 +146,21 @@ def fill_trade_embed_fields(
         # to do this, we add a 3rd empty field on each line (since 3 fields per line)
         i = 1
         while i < len(trader1_proposal) or i < len(trader2_proposal):
-            embed.add_field(name="\u200B", value="\u200B", inline=True)  # empty
+            embed.add_field(name="\u200b", value="\u200b", inline=True)  # empty
 
             if i < len(trader1_proposal):
-                embed.add_field(name="\u200B", value=trader1_proposal[i], inline=True)
+                embed.add_field(name="\u200b", value=trader1_proposal[i], inline=True)
             else:
-                embed.add_field(name="\u200B", value="\u200B", inline=True)
+                embed.add_field(name="\u200b", value="\u200b", inline=True)
 
             if i < len(trader2_proposal):
-                embed.add_field(name="\u200B", value=trader2_proposal[i], inline=True)
+                embed.add_field(name="\u200b", value=trader2_proposal[i], inline=True)
             else:
-                embed.add_field(name="\u200B", value="\u200B", inline=True)
+                embed.add_field(name="\u200b", value="\u200b", inline=True)
             i += 1
 
         # always add an empty field at the end, otherwise the alignment is off
-        embed.add_field(name="\u200B", value="\u200B", inline=True)
+        embed.add_field(name="\u200b", value="\u200b", inline=True)
 
     if len(embed) > 6000:
         if not compact:
